% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_mortality_age_range}
\alias{get_mortality_age_range}
\title{get_mortality_age_range}
\usage{
get_mortality_age_range(age)
}
\arguments{
\item{age}{age range code provided by Secretaria de Salud dictionary 'edad_agru'}
}
\description{
Returns the mortality registers by age range from years 2012-2020
}
