% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_mortality_age_range_year}
\alias{get_mortality_age_range_year}
\title{get_mortality_age_range_year}
\usage{
get_mortality_age_range_year(age, year)
}
\arguments{
\item{age}{age range provided by Secretaria de Salud dictionary 'edad_agru'}

\item{year}{year from 2012 - 2020 by Secretaria de Salud dictionary 'anio_regis'}
}
\description{
Returns the mortality registers by age range and specific year (2012 - 2020)
}
