% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_mortality_medical_year}
\alias{get_mortality_medical_year}
\title{get_mortality_medical_year}
\usage{
get_mortality_medical_year(sex, year)
}
\arguments{
\item{medical}{medical code provided by Secretaria de Salud dictionary 'asist_medi'}

\item{state}{State code provided by Secretaria de Salud dictionary 'ent_regis'}
}
\description{
Returns the mortality registers by medical attention received and specific state entity
}
