% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_mortality_scholarship_state}
\alias{get_mortality_scholarship_state}
\title{get_mortality_scholarship_state}
\usage{
get_mortality_scholarship_state(scholarship, state)
}
\arguments{
\item{scholarship}{Scholarship code provided by Secretaria de Salud dictionary 'escolarida'}

\item{state}{State code provided by Secretaria de Salud dictionary 'ent_regis'}
}
\description{
This function return the mortality of an specific state of Mexico with the
selected scholarship
}
