% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_mortality_scholarship_year}
\alias{get_mortality_scholarship_year}
\title{get_mortality_scholarship_year}
\usage{
get_mortality_scholarship_year(scholarship, year)
}
\arguments{
\item{scholarship}{Scholarship code provided by Secretaria de Salud dictionary 'escolarida'}

\item{year}{The desired year of the mortality registers 'anio_regis'}
}
\value{
A data frame with the data
}
\description{
This function return the mortality of Mexico of a specific year and only
of a certain scholarship
}
