% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_mortality_sex_state}
\alias{get_mortality_sex_state}
\title{get_mortality_sex_state}
\usage{
get_mortality_sex_state(sex, state)
}
\arguments{
\item{sex}{sex code provided by Secretaria de Salud dictionary 'sexo'}

\item{state}{State code provided by Secretaria de Salud dictionary 'ent_regis'}
}
\description{
Returns the mortality registers by sex type and specific state from Mexico
}
