% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_mortality_sex_year}
\alias{get_mortality_sex_year}
\title{get_mortality_sex_year}
\usage{
get_mortality_sex_year(sex, year)
}
\arguments{
\item{sex}{sex code provided by Secretaria de Salud dictionary 'sexo'}

\item{year}{State code provided by Secretaria de Salud dictionary 'anio_regis'}
}
\description{
This function returns the mortality registers based on sex type and year
}
