% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_mortality_state}
\alias{get_mortality_state}
\title{get_mortality_state}
\usage{
get_mortality_state(state)
}
\arguments{
\item{state}{State code provided by Secretaria de Salud dictionary 'ent_regis'}
}
\value{
A data frame with the data
}
\description{
This function return the mortality registers of a State from the selected state
from 2012-2020
}
