% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_mortality_state_year}
\alias{get_mortality_state_year}
\title{get_mortality_state_year}
\usage{
get_mortality_state_year(state, year)
}
\arguments{
\item{state}{State code provided by Secretaria de Salud dictionary 'ent_regis'}

\item{year}{The desired year of the mortality registers 'anio_regis'}
}
\value{
A data frame with the data
}
\description{
This function return the mortality registers of a State from the selected state
and the selected year
}
