% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_top_morbidity}
\alias{get_top_morbidity}
\title{get_top_morbidity}
\usage{
get_top_morbidity()
}
\description{
Returns the mortality registers by top ten specific dicease  code provided
by Secretaria de Salud dictionary 'lista_mex' from years 2012 -2020
}
