% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_top_morbidity_year}
\alias{get_top_morbidity_year}
\title{get_top_morbidity_year}
\usage{
get_top_morbidity_year(year)
}
\arguments{
\item{year}{year from 2012 - 2020 by Secretaria de Salud dictionary 'anio_regis'}
}
\description{
Returns the mortality registers by top ten specific dicease code provided by
Secretaria de Salud dictionary 'lista_mex' and specific year(2012 -2020)
}
