% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_top_mortality_states}
\alias{get_top_mortality_states}
\title{get_top_mortality_states}
\usage{
get_top_mortality_states()
}
\description{
Returns the mortality registers by top ten specific states codes provided by
Secretaria de Salud dictionary 'ent_regis' and years 2012 - 2020
}
