% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_top_mortality_states_year}
\alias{get_top_mortality_states_year}
\title{get_top_mortality_states_year}
\usage{
get_top_mortality_states_year(year)
}
\arguments{
\item{year}{year from 2012 - 2020 by Secretaria de Salud dictionary 'anio_regis'}
}
\description{
Returns the mortality registers by top ten states and specific year.
Codes provided by Secretaria de Salud dictionary 'ent_regis' and specific year(2012 -2020)
}
